Centralize job consumer logging configuration in logging_utils
Renamed _log_filters.py to logging_utils.py and added
configure_job_consumer_logging() utility function that:
- Adds filters to huey loggers to suppress/downgrade online scoring logs
- Sets alembic logger to WARNING level

Both _periodic_tasks_consumer.py and _huey_consumer.py now call this
utility function instead of duplicating the logging setup code.

SuppressOnlineScoringFilter downgrades INFO -> DEBUG for logs mentioning:
- run_online_trace_scorer
- run_online_session_scorer
- online_scoring_scheduler

Logs are shown only if logger's effective level is DEBUG or lower,
otherwise they're suppressed entirely.
